LLM_to_API/requirements.txt
langchain==0.3.3
langchain-google-genai 
requests
python-dotenv
numpy